Better JS error display & thumbnail handling
Allow the error message to show multiple lines. This means that several
errors can be shown when uploading a lot of files. Also show the file
that the error refers to

Update the thumbnail preview with the real thumbnail after uploading.
This fixes instances where the preview was generated after the file was
uploaded.

Allow an upload error to stop further uploading.